refactor: move ui to separate pkg
